+++ server/requirements.txt
@@ -1,4 +1,5 @@
 alembic==1.7.1
+bcrypt==3.2.0
 black==21.9b0
 click==8.0.1
 cryptography==3.4.8
